refactor(api): split 6 remaining large files under 700-line guardrail
Extract response constants, Pydantic models, and shared helpers into
companion modules for: auth, monitoring_watchlists, risk/alerts,
indicator_cache, governance_dashboard, watchlist.

All 14 files pass syntax validation. No API behavior changed.
